[enigma2] Fix compiler warnings across multiple architectures (arm, mipsel, aarch64) - Fix implicit-fallthrough warnings with proper fall-through comments - Fix format-overflow warnings by using appropriately sized buffers - Fix sign-compare warnings with proper type casts - Fix unused-variable warnings by removing dead code (scan.cpp, elistbox.cpp) - Fix macro-redefined warnings with #ifndef guards (esize.h, python.h) - Guard DREAMBOX_DUAL_TUNER-only functions with #ifdef (dvbci.cpp) - Add -Wno-psabi to suppress ARM ABI change notes for STL iterator passing
cherry-picked > https://github.com/openatv/enigma2/commit/7757bff33aaccd9a3ca3d6c040fc9ae8f798805a

+++ lib/service/servicedvbsoftdecoder.cpp
@@ -429,7 +429,7 @@ void eDVBSoftDecoder::streamHealthCheck()
 		int64_t elapsed = now - m_last_health_check;
 		if (elapsed > 2000)
 		{
-			eDebug("[eDVBSoftDecoder] MainLoop was blocked for %lldms, skipping stall check", elapsed);
+			eDebug("[eDVBSoftDecoder] MainLoop was blocked for %lldms, skipping stall check", (long long)elapsed);
 			m_stall_count = 0;
 			m_stream_stalled = false;
 			m_last_pts = 0;
